debug HF1

+++ Common/record_results.py
@@ -4,6 +4,7 @@ import csv
 import json
 from collections import namedtuple
 from Common import rename_file
+import Common
 
 
 def get_init_distance(job):
@@ -22,20 +23,23 @@ def creatcsv(path, cal_type):
     headers = [
         'displacement',
         'distance(A)',
-        'E(au)',
-        'Eupperlayer(au)',
-        'Eunderlayer(au)']
-    with open(csv_path, 'w', newline='') as f:
-        f_csv = csv.writer(f)
-        f_csv.writerow(headers)
+        'E',
+        'Eupperlayer',
+        'Eunderlayer',
+        'Einterlayer']
+    if not os.path.exists(csv_path):
+        with open(csv_path, 'w', newline='') as f:
+            f_csv = csv.writer(f)
+            f_csv.writerow(headers)
 
 
 def creat_json_file(path, cal_type):
     json_file = os.path.join(path, cal_type)
     json_file = os.path.join(json_file, '{}.json'.format(cal_type))
-    data = {}
-    with open(json_file, 'w') as f:
-        json.dump(data, f, indent=4)
+    if not os.path.exists(json_file):
+        data = {}
+        with open(json_file, 'w') as f:
+            json.dump(data, f, indent=4)
 
 
 def data_saving(path, line, cal_type):
@@ -55,19 +59,18 @@ def read_result(job, energy_func, init_distance):
     z = z + init_distance
     z = round(z, 2)
     z = str(z)
-    layertype = job.layertype
     x_z = namedtuple('x_z', ['x', 'z'])
     x_and_z = x_z(x, z)
     if job.method != 'lpm2':
         energy = energy_func(job)
         return x_and_z, energy
     else:
+        # should correct then
         Elmp2, Escs_lmp2 = energy_func(path)
         return x_and_z, Elmp2, Escs_lmp2
 
 
 def read_all_results(jobs, cal_type, energy_func, init_distance=None):
-    readed_jobs = []
 
     # get the initial distance of the system
     if init_distance is None:
@@ -90,8 +93,9 @@ def read_all_results(jobs, cal_type, energy_func, init_distance=None):
                     continue
 
     path = jobs[0].root_path
-    creatcsv(path, cal_type)
     creat_json_file(path, cal_type)
+    json_file = os.path.join(path, cal_type)
+    json_file = os.path.join(json_file, '{}.json'.format(cal_type))
 
     bilayer = {}
     upperlayer = {}
@@ -106,35 +110,49 @@ def read_all_results(jobs, cal_type, energy_func, init_distance=None):
         elif layertype == 'upperlayer':
             upperlayer[x_and_z] = energy
         # record to json file
-        json_file = os.path.join(path, cal_type)
-        json_file = os.path.join(json_file, '{}.json'.format(cal_type))
         record_to_json(jobs[i], json_file, energy, layertype, init_distance)
 
+    # calculation of layer energy
+    layer_energy_dict = {}
+    if len(bilayer) > 0 and len(underlayer) > 0:
+        for job in jobs:
+            if job.layertype == 'bilayer':
+                x_and_z, _ = read_result(job, energy_func, init_distance)
+                if x_and_z in bilayer and x_and_z in underlayer and x_and_z in upperlayer:
+                    layer_energy = Common.cal_layer_energy(bilayer[x_and_z], upperlayer[x_and_z], underlayer[x_and_z])
+                    layer_energy_dict[x_and_z] = layer_energy
+                    record_to_json(job, json_file, layer_energy, 'layer_energy', init_distance)
+
+    # record to .csv file
     lines = []
     for x_and_z, energy in bilayer.items():
         line = []
         line.append(x_and_z.x)
         line.append(x_and_z.z)
-        line.append(energy)
+        line.append(energy[0])
         if x_and_z in upperlayer:
-            line.append(upperlayer[x_and_z])
+            line.append(upperlayer[x_and_z][0])
         else:
             line.append('NaN')
         if x_and_z in underlayer:
-            line.append(underlayer[x_and_z])
+            line.append(underlayer[x_and_z][0])
+        else:
+            line.append('NaN')
+        if x_and_z in layer_energy_dict:
+            line.append(layer_energy_dict[x_and_z][0])
         else:
             line.append('NaN')
         lines.append(line)
 
     csv_path = os.path.join(path, '{}.csv'.format(cal_type))
     rename_file(path, '{}.csv'.format(cal_type))
+    creatcsv(path, cal_type)
     with open(csv_path, 'a', newline='') as f:
         f_csv = csv.writer(f)
         f_csv.writerows(lines)
 
 
 def record_to_json(job, json_file, energy, layertype='bilayer', init_distance=3.1):
-    path = job.path
     coord = '{}'.format(job.coord)
     x = job.x
     if is_number(x):
@@ -146,6 +164,7 @@ def record_to_json(job, json_file, energy, layertype='bilayer', init_distance=3.
         data = json.load(f)
     if coord not in data:
         data[coord] = {}
+    path = job.path
     data[coord][layertype] = {
         'energy': energy,
         'x': x,
